feat: efficient HMR updates
This PR introduces new React hooks for subscribing to hot-reloadable data provided by the "vite-plugin-react-pages/dist/client" module. These hooks include:
  - useStaticData: Subscribe to the `.staticData` of a specific page or all pages.
  - usePagePaths: Subscribe to the array of paths for all pages.
  - usePageModule: Subscribe to the module referred to by the `dataPath` of a specific page.
  - useTheme: Subscribe to the `Theme` component.

In most cases, `useStaticData` is the only hook used by the application/theme. It replaces the `staticData` prop previously passed to the `Theme` component. The idea is to depend on static data where it's needed, instead of having to re-render the entire application when the static data is updated.

By calling `useStaticData()`, the component subscribes to **all static data.** By calling `useStaticData(path)`, the component only re-renders when the static data of that specific path is updated. You can even pass a selector if you only need a specific piece of static data. In the example below, the caller only re-renders when the page title is changed.

```ts
const title = useStaticData(path, data => data.main.title)
```

The other hooks are mostly for the internal infrastructure of react-pages.

`useTheme` is used by the `PageLoader` component (which is now wrapped with `React.memo` to avoid re-rendering when a page is added or removed).

`usePagePaths` is used by the `App` component for updating the page routes when a page is added or removed.

`usePageModule` is used by the `useAppState` hook (which I've rewritten) for reloading a page when its `dataPath` is changed. It even uses the `/404` page if a path does not exist. Once the path *does* exist, HMR will re-render the `PageLoader` automatically with a new `usePageModule` promise.

How does this all work?
---

When HMR is enabled, we use a library called [jotai](https://npmjs.org/jotai), which lets us re-render components only when the data they depend on is changed. The `client/state.ts` module (with access to our Jotai atoms) plugs into Vite's HMR API. When the `@!virtual-modules/pages` module is updated, we diff it with its previous module to determine what has changed (see `setPagesAtom` in `client/state.ts`).

In production, all that logic is tree-shaked, and we're left with simple hooks that return immutable data.